+++ javascript/stimulus_reflex.js
@@ -1,10 +1,11 @@
 import { Controller } from 'stimulus'
 import CableReady from 'cable_ready'
+import serializeForm from 'form-serialize'
 import { defaultSchema } from './schema'
 import { getConsumer } from './consumer'
 import { dispatchLifecycleEvent } from './lifecycle'
 import { allReflexControllers } from './controllers'
-import { uuidv4 } from './utils'
+import { uuidv4, debounce } from './utils'
 import Log from './log'
 import {
   attributeValue,
@@ -107,7 +108,6 @@ const extendStimulusController = controller => {
         reflexId: reflexId
       }
       const { subscription } = this.StimulusReflex
-      const { connection } = subscription.consumer
 
       if (!this.isActionCableConnectionOpen())
         throw 'The ActionCable connection is not open! `this.isActionCableConnectionOpen()` must return true before calling `this.stimulate()`'
@@ -118,7 +118,21 @@ const extendStimulusController = controller => {
 
       dispatchLifecycleEvent('before', element)
 
-      subscription.send(data)
+      setTimeout(() => {
+        const { params } = element.reflexData || {}
+        element.reflexData = {
+          ...data,
+          params: {
+            ...params,
+            ...serializeForm(element.closest('form'), {
+              hash: true,
+              empty: true
+            })
+          }
+        }
+
+        subscription.send(element.reflexData)
+      })
 
       if (debugging) {
         Log.request(
@@ -189,9 +203,9 @@ class StimulusReflexController extends Controller {
 }
 
 // Sets up declarative reflex behavior.
-// Any elements that define data-reflex will automatcially be wired up with the default StimulusReflexController.
+// Any elements that define data-reflex will automatically be wired up with the default StimulusReflexController.
 //
-const setupDeclarativeReflexes = () => {
+const setupDeclarativeReflexes = debounce(() => {
   document
     .querySelectorAll(`[${stimulusApplication.schema.reflexAttribute}]`)
     .forEach(element => {
@@ -234,7 +248,7 @@ const setupDeclarativeReflexes = () => {
           actionValue
         )
     })
-}
+}, 20)
 
 // compute the DOM element(s) which will be the morph root
 // use the data-reflex-root attribute on the reflex or the controller
@@ -287,16 +301,17 @@ const initialize = (application, initializeOptions = {}) => {
 
 if (!document.stimulusReflexInitialized) {
   document.stimulusReflexInitialized = true
-  window.addEventListener('load', () => setTimeout(setupDeclarativeReflexes, 1))
-  document.addEventListener('turbolinks:load', () =>
-    setTimeout(setupDeclarativeReflexes, 1)
-  )
-  document.addEventListener('cable-ready:after-morph', () =>
-    setTimeout(setupDeclarativeReflexes, 1)
-  )
-  document.addEventListener('ajax:complete', () =>
-    setTimeout(setupDeclarativeReflexes, 1)
-  )
+
+  window.addEventListener('load', () => {
+    setupDeclarativeReflexes()
+    const observer = new MutationObserver(setupDeclarativeReflexes)
+    observer.observe(document.documentElement, {
+      attributes: true,
+      childList: true,
+      subtree: true
+    })
+  })
+
   // Trigger success and after lifecycle methods from before-morph to ensure we can find a reference
   // to the source element in case it gets removed from the DOM via morph.
   // This is safe because the server side reflex completed successfully.
@@ -326,34 +341,57 @@ if (!document.stimulusReflexInitialized) {
     dispatchLifecycleEvent('success', element)
     if (debugging) Log.success(response)
   })
-  document.addEventListener('stimulus-reflex:500', event => {
-    const { reflexId, attrs, error } = event.detail.stimulusReflex || {}
+  document.addEventListener('stimulus-reflex:server-message', event => {
+    const { reflexId, attrs, serverMessage } = event.detail.stimulusReflex || {}
+    const { subject, body } = serverMessage
     const element = findElement(attrs)
     const promise = promises[reflexId]
+    const subjects = {
+      error: true,
+      halted: true
+    }
 
-    if (element) element.reflexError = error
+    if (element && subject == 'error') element.reflexError = body
 
     const response = {
       data: promise && promise.data,
       element,
       event,
-      toString: () => error
+      events: promise && promise.events,
+      toString: () => body
     }
 
     if (promise) {
       delete promises[reflexId]
-      promise.reject(response)
+
+      if (subject == 'error') {
+        promise.reject(response)
+      } else {
+        promise.resolve(response)
+      }
     }
 
-    dispatchLifecycleEvent('error', element)
-    if (debugging) Log.error(response)
+    if (element && subjects[subject]) dispatchLifecycleEvent(subject, element)
+
+    if (debugging) {
+      switch (subject) {
+        case 'error':
+          Log.error(response)
+          break
+        case 'halted':
+          Log.success(response, { halted: true })
+          break
+        default:
+          Log.success(response)
+          break
+      }
+    }
   })
 }
 
 export default {
   initialize,
   register,
-  setupDeclarativeReflexes,
   get debug () {
     return debugging
   },