+++ javascript/utils.js
@@ -21,3 +21,14 @@ export const camelize = (value, uppercaseFirstLetter = true) => {
 
   return value
 }
+
+export const debounce = (callback, delay = 250) => {
+  let timeoutId
+  return (...args) => {
+    clearTimeout(timeoutId)
+    timeoutId = setTimeout(() => {
+      timeoutId = null
+      callback(...args)
+    }, delay)
+  }
+}